Sentences generation on faqs
