+++ bot/admin/web/src/app/faq/faq-management/faq-management-edit/faq-management-edit.component.html
@@ -181,12 +181,11 @@
               nbTooltip="Generate sentences"
               ghost
               shape="round"
+              status="primary"
               (click)="generateSentences()"
+              [disabled]="!utterances.controls.length"
             >
-              <nb-icon
-                icon="bulb-outline"
-                status="primary"
-              ></nb-icon>
+              <nb-icon icon="bulb-outline"></nb-icon>
             </button>
           </div>
         </tock-form-control>
