+++ bot/admin/web/src/app/shared/components/sentences-generation/models/completion.ts
