+++ bot/admin/web/src/app/shared/components/sentences-generation/models/index.ts
