+++ bot/admin/web/src/app/shared/components/sentences-generation/models/sentencesGenerationOptions.ts
@@ -3,6 +3,5 @@ export interface SentencesGenerationOptions {
   smsLanguage: boolean;
   abbreviatedLanguage: boolean;
   temperature: number;
-  sentenceExample: string;
   sentencesExample: string[];
 }
