+++ bot/admin/web/src/app/shared/components/sentences-generation/sentences-generation-list/sentences-generation-list.component.spec.ts
