+++ bot/admin/web/src/app/shared/components/sentences-generation/sentences-generation-options/sentences-generation-options.component.spec.ts
