+++ bot/admin/web/src/app/shared/components/sentences-generation/sentences-generation.component.html
@@ -1,0 +1,88 @@
+<nb-card [nbSpinner]="loading">
+  <nb-card-header class="d-flex justify-content-between align-items-start gap-1">
+    Sentences generation
+    <button
+      type="button"
+      nbButton
+      shape="round"
+      ghost
+      nbTooltip="Close"
+      (click)="dialogRef.close('cancel')"
+    >
+      <nb-icon icon="close-outline"></nb-icon>
+    </button>
+  </nb-card-header>
+
+  <nb-card-body>
+    <section *ngIf="llmNotEnabled">
+      <div class="text-warning font-weight-bold">LLM settings are missing.</div>
+
+      <span *ngIf="!state.hasRole(UserRole.admin)"> Contact an administrator </span>
+
+      <button
+        *ngIf="state.hasRole(UserRole.admin)"
+        nbButton
+        status="primary"
+        class="mt-3"
+        (click)="jumpToLlmSettings()"
+      >
+        LLM settings
+      </button>
+    </section>
+
+    <section *ngIf="!llmNotEnabled">
+      <nb-alert
+        *ngIf="informNoResult"
+        status="warning"
+        closable
+        (click)="informNoResult = false"
+        class="pointer"
+        >The query returned no results. Try restarting the request or selecting other example sentences.</nb-alert
+      >
+
+      <tock-sentences-generation-options
+        *ngIf="!generatedSentences.length"
+        [options]="options"
+        [sentences]="sentences"
+        (onOptionsUpdate)="updateOptions($event)"
+      ></tock-sentences-generation-options>
+
+      <tock-sentences-generation-list
+        #sentencesGenerationListComp
+        *ngIf="generatedSentences.length"
+        [generatedSentences]="generatedSentences"
+        (onBackOptions)="backOptions()"
+        (onGenerate)="generate()"
+      ></tock-sentences-generation-list>
+    </section>
+  </nb-card-body>
+
+  <nb-card-footer
+    class="card-footer-actions"
+    *ngIf="!llmNotEnabled"
+  >
+    <button
+      *ngIf="!generatedSentences.length"
+      type="submit"
+      nbButton
+      status="primary"
+      size="small"
+      [disabled]="!options.sentencesExample?.length"
+      (click)="generate()"
+    >
+      Generate
+    </button>
+
+    <button
+      *ngIf="generatedSentences.length"
+      type="submit"
+      nbButton
+      status="primary"
+      size="small"
+      [disabled]="!generatedSentencesSelection?.length"
+      (click)="validateSelection()"
+    >
+      Validate selection
+    </button>
+  </nb-card-footer>
+</nb-card>
