+++ bot/admin/web/src/app/shared/components/sentences-generation/sentences-generation.component.scss
@@ -1,4 +1,6 @@
 nb-card {
+  width: 90vw;
+  max-width: 90vw;
   max-height: 90vh;
   margin: 0;
 }
