--- bot/admin/web/src/app/shared/modules/index.ts
@@ -1,1 +1,0 @@
-export * from './sentences-generation/sentences-generation.module';
