--- bot/admin/web/src/app/shared/modules/sentences-generation/components/index.ts
@@ -1,4 +1,0 @@
-export * from './sentences-generation-content/sentences-generation-content.component';
-export * from './sentences-generation-list/sentences-generation-list.component';
-export * from './sentences-generation-options/sentences-generation-options.component';
-export * from './sentences-generation-wrapper/sentences-generation-wrapper.component';
