--- bot/admin/web/src/app/shared/modules/sentences-generation/components/sentences-generation-content/sentences-generation-content.component.html
@@ -1,48 +1,0 @@
-<div
-  *ngIf="closable || closable === ''"
-  class="d-flex justify-content-between align-items-center font-weight-bold mb-3"
->
-  Sentences generation
-  <button
-    nbButton
-    type="button"
-    nbTooltip="Close generate sentences"
-    ghost
-    shape="round"
-    (click)="close()"
-  >
-    <nb-icon icon="close-outline"></nb-icon>
-  </button>
-</div>
-
-<nb-alert
-  *ngIf="showAlert"
-  outline="warning"
-  closable
-  (close)="closeAlert()"
->
-  <div class="d-flex align-items-center gap-1">
-    <nb-icon
-      icon="alert-triangle-outline"
-      status="warning"
-    ></nb-icon>
-    Please note that data will be transmitted to an external partner. Be careful not to send sentences containing "sensitive" information.
-  </div>
-</nb-alert>
-
-<section>
-  <tock-sentences-generation-options
-    *ngIf="!generatedSentences.length"
-    [options]="options"
-    [sentences]="sentences"
-    (onGenerate)="generate($event)"
-  ></tock-sentences-generation-options>
-
-  <tock-sentences-generation-list
-    *ngIf="generatedSentences.length"
-    [generatedSentences]="generatedSentences"
-    (onBackOptions)="backOptions()"
-    (onGenerate)="generate()"
-    (onValidateSelection)="validateSelection($event)"
-  ></tock-sentences-generation-list>
-</section>
