--- bot/admin/web/src/app/shared/modules/sentences-generation/components/sentences-generation-wrapper/sentences-generation-wrapper.component.html
@@ -1,41 +1,0 @@
-<nb-card [nbSpinner]="loading">
-  <nb-card-header class="d-flex justify-content-between align-items-start gap-1">
-    Sentences generation
-    <button
-      type="button"
-      nbButton
-      shape="round"
-      ghost
-      nbTooltip="Close"
-      (click)="dialogRef.close('cancel')"
-    >
-      <nb-icon icon="close-outline"></nb-icon>
-    </button>
-  </nb-card-header>
-
-  <nb-card-body>
-    <div *ngIf="llmSettingsMissing">
-      <div class="text-warning font-weight-bold">LLM settings are missing.</div>
-
-      <span *ngIf="!state.hasRole(UserRole.admin)"> Contact an administrator </span>
-
-      <button
-        *ngIf="state.hasRole(UserRole.admin)"
-        nbButton
-        status="primary"
-        class="mt-3"
-        (click)="jumpToLlmSettings()"
-      >
-        LLM settings
-      </button>
-    </div>
-
-    <tock-sentences-generation-content
-      *ngIf="!llmSettingsMissing"
-      [sentences]="sentences"
-      [errors]="errors"
-      (onLoading)="handleLoading($event)"
-      (onValidateSelection)="handleGeneratedSentences($event)"
-    ></tock-sentences-generation-content>
-  </nb-card-body>
-</nb-card>
