--- bot/admin/web/src/app/shared/modules/sentences-generation/services/index.ts
@@ -1,2 +1,0 @@
-export * from './sentences-generation/sentences-generation.service';
-export * from './sentences-generation-api/sentences-generation-api.service';
