--- bot/admin/web/src/app/shared/modules/sentences-generation/services/sentences-generation-api/sentences-generation-api.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SentencesGenerationApiService } from './sentences-generation-api.service';
-
-describe('SentencesGenerationApiService', () => {
-  let service: SentencesGenerationApiService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SentencesGenerationApiService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
